HM-06 Moved the service interface
