--- src/main/java/br/com/ead/home/services/Service.java
@@ -1,4 +1,0 @@
-package br.com.ead.home.services;
-
-public interface Service {
-}
