+++ src/main/java/br/com/ead/home/services/api/BookablePreferenceService.java
@@ -2,7 +2,6 @@ package br.com.ead.home.services.api;
 
 import br.com.ead.home.models.api.TimeSlotPreferences;
 import br.com.ead.home.models.primitives.ClinicianId;
-import br.com.ead.home.services.Service;
 
 public interface BookablePreferenceService extends Service {
