+++ src/main/java/br/com/ead/home/services/api/ScheduleAvailabilityService.java
@@ -2,7 +2,6 @@ package br.com.ead.home.services.api;
 
 import br.com.ead.home.models.api.TimeSlot;
 import br.com.ead.home.models.primitives.ClinicianId;
-import br.com.ead.home.services.Service;
 
 import java.time.ZonedDateTime;
 import java.util.Set;
