+++ src/main/java/br/com/ead/home/services/api/Service.java
@@ -1,0 +1,4 @@
+package br.com.ead.home.services.api;
+
+public interface Service {
+}
