+++ src/main/java/br/com/ead/home/services/delegates/lookups/ServiceLookup.java
@@ -1,6 +1,6 @@
 package br.com.ead.home.services.delegates.lookups;
 
-import br.com.ead.home.services.Service;
+import br.com.ead.home.services.api.Service;
 import br.com.ead.home.services.delegates.types.ServicePartitionType;
 import br.com.ead.home.services.delegates.types.ServiceStageType;
